pyproject.toml
[project]
name = "green-wind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ruff>=0.6.1",
    "fastapi[standard]>=0.112.1",
    "uvicorn>=0.30.6",
    "polyfactory>=2.16.2",
    "sqlmodel>=0.0.21",
    "pydantic-settings>=2.4.0",
    "psycopg2>=2.9.9",
    "ipykernel>=6.29.5",
    "pandas>=2.2.2",
    "polars>=1.5.0",
    "scipy>=1.14.1",
    "matplotlib>=3.9.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
